Fixed type size issue exposed by CI

+++ src/binary80/exp/tint.h
@@ -378,7 +378,7 @@ tint_tod (const tint_t *a, uint64_t err)
     if (a->ex < -1074) // |a| < 2^-1075
       return (a->sgn ? -0x1p-1074 : 0x1p-1074) * 0.5;
     // 2^-1075 <= |a| < 2^-1074
-    int mid = a->h == (1ull << 63) && a->m == 0 && a->l == 0;
+    int64_t mid = a->h == (1ull << 63) && a->m == 0 && a->l == 0;
     // if mid, |a| = 2^-1075
     return (a->sgn ? -0x1p-1074 : 0x1p-1074) * (mid ? 0.5 : 0.75);
   }
